Implement migration (#120)
* Make easier to copy paste to execute

* Make migration work

* Bump edit_note action schema to v3

* Use UnixNano for timstamps

* Fix schema

+++ core/reducer.go
@@ -26,6 +26,8 @@ func ReduceAll(ctx infra.DnoteCtx, tx *sql.Tx, actionSlice []actions.Action) err
 // Reduce transitions the local dnote state by consuming the action returned
 // from the server
 func Reduce(ctx infra.DnoteCtx, tx *sql.Tx, action actions.Action) error {
+	log.Debug("reducing %s. uuid: %s, schema: %d, timestamp: %d\n", action.Type, action.UUID, action.Schema, action.Timestamp)
+
 	var err error
 
 	switch action.Type {
@@ -63,12 +65,16 @@ func getBookUUIDWithTx(tx *sql.Tx, bookLabel string) (string, error) {
 }
 
 func handleAddNote(ctx infra.DnoteCtx, tx *sql.Tx, action actions.Action) error {
-	var data actions.AddNoteDataV1
+	if action.Schema != 2 {
+		return errors.Errorf("data schema '%d' not supported", action.Schema)
+	}
+
+	var data actions.AddNoteDataV2
 	if err := json.Unmarshal(action.Data, &data); err != nil {
 		return errors.Wrap(err, "parsing the action data")
 	}
 
-	log.Debug("reducing add_note. action: %+v. data: %+v\n", action, data)
+	log.Debug("data: %+v\n", data)
 
 	bookUUID, err := getBookUUIDWithTx(tx, data.BookName)
 	if err != nil {
@@ -76,8 +82,9 @@ func handleAddNote(ctx infra.DnoteCtx, tx *sql.Tx, action actions.Action) error 
 	}
 
 	var noteCount int
-	err = tx.QueryRow("SELECT count(uuid) FROM notes WHERE uuid = ? AND book_uuid = ?", data.NoteUUID, bookUUID).Scan(&noteCount)
-	if err != nil {
+	if err := tx.
+		QueryRow("SELECT count(uuid) FROM notes WHERE uuid = ? AND book_uuid = ?", data.NoteUUID, bookUUID).
+		Scan(&noteCount); err != nil {
 		return errors.Wrap(err, "counting note")
 	}
 
@@ -91,7 +98,7 @@ func handleAddNote(ctx infra.DnoteCtx, tx *sql.Tx, action actions.Action) error 
 
 	_, err = tx.Exec(`INSERT INTO notes
 	(uuid, book_uuid, content, added_on, public)
-	VALUES (?, ?, ?, ?, ?)`, data.NoteUUID, bookUUID, data.Content, action.Timestamp, false)
+	VALUES (?, ?, ?, ?, ?)`, data.NoteUUID, bookUUID, data.Content, action.Timestamp, data.Public)
 	if err != nil {
 		return errors.Wrap(err, "inserting a note")
 	}
@@ -100,12 +107,16 @@ func handleAddNote(ctx infra.DnoteCtx, tx *sql.Tx, action actions.Action) error 
 }
 
 func handleRemoveNote(ctx infra.DnoteCtx, tx *sql.Tx, action actions.Action) error {
-	var data actions.RemoveNoteDataV1
+	if action.Schema != 2 {
+		return errors.Errorf("data schema '%d' not supported", action.Schema)
+	}
+
+	var data actions.RemoveNoteDataV2
 	if err := json.Unmarshal(action.Data, &data); err != nil {
 		return errors.Wrap(err, "parsing the action data")
 	}
 
-	log.Debug("reducing remove_note. action: %+v. data: %+v\n", action, data)
+	log.Debug("data: %+v\n", data)
 
 	_, err := tx.Exec("DELETE FROM notes WHERE uuid = ?", data.NoteUUID)
 	if err != nil {
@@ -115,7 +126,7 @@ func handleRemoveNote(ctx infra.DnoteCtx, tx *sql.Tx, action actions.Action) err
 	return nil
 }
 
-func buildEditNoteQuery(ctx infra.DnoteCtx, tx *sql.Tx, noteUUID, bookUUID string, ts int64, data actions.EditNoteDataV2) (string, []interface{}, error) {
+func buildEditNoteQuery(ctx infra.DnoteCtx, tx *sql.Tx, noteUUID string, ts int64, data actions.EditNoteDataV3) (string, []interface{}, error) {
 	setTmpl := "edited_on = ?"
 	queryArgs := []interface{}{ts}
 
@@ -127,37 +138,37 @@ func buildEditNoteQuery(ctx infra.DnoteCtx, tx *sql.Tx, noteUUID, bookUUID strin
 		setTmpl = fmt.Sprintf("%s, public = ?", setTmpl)
 		queryArgs = append(queryArgs, *data.Public)
 	}
-	if data.ToBook != nil {
-		bookUUID, err := getBookUUIDWithTx(tx, *data.ToBook)
+	if data.BookName != nil {
+		setTmpl = fmt.Sprintf("%s, book_uuid = ?", setTmpl)
+
+		bookUUID, err := getBookUUIDWithTx(tx, *data.BookName)
 		if err != nil {
-			return "", []interface{}{}, errors.Wrap(err, "getting destination book uuid")
+			return setTmpl, queryArgs, errors.Wrap(err, "getting book uuid")
 		}
 
-		setTmpl = fmt.Sprintf("%s, book_uuid = ?", setTmpl)
 		queryArgs = append(queryArgs, bookUUID)
 	}
 
-	queryTmpl := fmt.Sprintf("UPDATE notes SET %s WHERE uuid = ? AND book_uuid = ?", setTmpl)
-	queryArgs = append(queryArgs, noteUUID, bookUUID)
+	queryTmpl := fmt.Sprintf("UPDATE notes SET %s WHERE uuid = ?", setTmpl)
+	queryArgs = append(queryArgs, noteUUID)
 
 	return queryTmpl, queryArgs, nil
 }
 
 func handleEditNote(ctx infra.DnoteCtx, tx *sql.Tx, action actions.Action) error {
-	var data actions.EditNoteDataV2
+	if action.Schema != 3 {
+		return errors.Errorf("data schema '%d' not supported", action.Schema)
+	}
+
+	var data actions.EditNoteDataV3
 	err := json.Unmarshal(action.Data, &data)
 	if err != nil {
 		return errors.Wrap(err, "parsing the action data")
 	}
 
-	log.Debug("reducing edit_note v2. action: %+v. data: %+v\n", action, data)
+	log.Debug("data: %+v\n", data)
 
-	bookUUID, err := getBookUUIDWithTx(tx, data.FromBook)
-	if err != nil {
-		return errors.Wrap(err, "getting book uuid")
-	}
-
-	queryTmpl, queryArgs, err := buildEditNoteQuery(ctx, tx, data.NoteUUID, bookUUID, action.Timestamp, data)
+	queryTmpl, queryArgs, err := buildEditNoteQuery(ctx, tx, data.NoteUUID, action.Timestamp, data)
 	if err != nil {
 		return errors.Wrap(err, "building edit note query")
 	}
@@ -170,13 +181,17 @@ func handleEditNote(ctx infra.DnoteCtx, tx *sql.Tx, action actions.Action) error
 }
 
 func handleAddBook(ctx infra.DnoteCtx, tx *sql.Tx, action actions.Action) error {
+	if action.Schema != 1 {
+		return errors.Errorf("data schema '%d' not supported", action.Schema)
+	}
+
 	var data actions.AddBookDataV1
 	err := json.Unmarshal(action.Data, &data)
 	if err != nil {
 		return errors.Wrap(err, "parsing the action data")
 	}
 
-	log.Debug("reducing add_book. action: %+v. data: %+v\n", action, data)
+	log.Debug("data: %+v\n", data)
 
 	var bookCount int
 	err = tx.QueryRow("SELECT count(uuid) FROM books WHERE label = ?", data.BookName).Scan(&bookCount)
@@ -199,21 +214,33 @@ func handleAddBook(ctx infra.DnoteCtx, tx *sql.Tx, action actions.Action) error 
 }
 
 func handleRemoveBook(ctx infra.DnoteCtx, tx *sql.Tx, action actions.Action) error {
+	if action.Schema != 1 {
+		return errors.Errorf("data schema '%d' not supported", action.Schema)
+	}
+
 	var data actions.RemoveBookDataV1
 	if err := json.Unmarshal(action.Data, &data); err != nil {
 		return errors.Wrap(err, "parsing the action data")
 	}
 
-	log.Debug("reducing remove_book. action: %+v. data: %+v\n", action, data)
+	log.Debug("data: %+v\n", data)
 
-	var bookUUID string
-	err := tx.QueryRow("SELECT uuid FROM books WHERE label = ?", data.BookName).Scan(&bookUUID)
-	if err == sql.ErrNoRows {
+	var bookCount int
+	if err := tx.
+		QueryRow("SELECT count(uuid) FROM books WHERE label = ?", data.BookName).
+		Scan(&bookCount); err != nil {
+		return errors.Wrap(err, "counting note")
+	}
+
+	if bookCount == 0 {
 		// If book does not exist, another client added and removed the book, making the add_book action
 		// obsolete. noop.
 		return nil
-	} else if err != nil {
-		return errors.Wrap(err, "querying the book")
+	}
+
+	bookUUID, err := getBookUUIDWithTx(tx, data.BookName)
+	if err != nil {
+		return errors.Wrap(err, "getting book uuid")
 	}
 
 	_, err = tx.Exec("DELETE FROM notes WHERE book_uuid = ?", bookUUID)
